% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGislet_annot.R
\name{annot_from_gff3}
\alias{annot_from_gff3}
\title{Get annotation dataframe from gff3 format files}
\usage{
annot_from_gff3(
  CGislet,
  gff3,
  symbol_pattern = "Name=",
  ensembl_pattern = "gene_id=",
  description = T,
  description_pattern = "description="
)
}
\arguments{
\item{CGislet}{CpG islet object}

\item{gff3}{gff3 file path of data.frame}

\item{symbol_pattern}{character;prefix of SYMBOL in annotation file}

\item{ensembl_pattern}{character;prefix of ENSEMBL names in annotation file}

\item{description}{logical;if T, return dataframe contain description column}

\item{description_pattern}{character;prefix of gene descriptions in annotation file}
}
\value{
annotation dataframe includes ENSEMBL, SYMBOL (and descriptions)
}
\description{
Get annotation dataframe from gff3 format files
}
