% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGislet.R
\name{CG_islet_structure}
\alias{CG_islet_structure}
\title{Identification of the structure of CG_islet}
\usage{
CG_islet_structure(CG_site, min_core_site = 5, min_linker_site = 5)
}
\arguments{
\item{min_core_site}{Minimum number of core sites per core region.}

\item{min_linker_site}{Minimum number of linker sites per linker region.}

\item{CG_islet}{CG_site object containing CG_islet and CG_islet_CG_levels column.}
}
\value{
CG_islet object include structure of CG_islet.
}
\description{
Identification of the structure of CG_islet
}
