% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGislet.R
\name{CG_levels_check}
\alias{CG_levels_check}
\title{Correct the structure of CG_islet}
\usage{
CG_levels_check(CG_levels, min_core_site = 5, min_linker_site = 5)
}
\arguments{
\item{CG_levels}{CG_levels of each CG site (1:core,2:flank)}

\item{min_core_site}{Minimum number of core sites per core region}

\item{min_linker_site}{Minimum number of linker sites per linker region}
}
\description{
Correct the structure of CG_islet
}
