% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGislet.R
\name{CG_site_pos}
\alias{CG_site_pos}
\title{Get CpG positions in each CGislet region}
\usage{
CG_site_pos(CGislet, CG_site)
}
\arguments{
\item{CGislet}{CGislet objects contain CGislet_type column}

\item{CpG_site}{CG_site objects contain CG_islet column}
}
\value{
CG_site with CpG positions
}
\description{
Get CpG positions in each CGislet region
}
