% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGislet.R
\name{CG_site_pos_detect}
\alias{CG_site_pos_detect}
\title{Get CG positions in each CGislet region}
\usage{
CG_site_pos_detect(CpG_site)
}
\arguments{
\item{CpG_site}{CG_site objects contain CG_islet column}
}
\value{
CG_site objects contain number, site_pos and site_pos_type columns
}
\description{
Get CG positions in each CGislet region
}
