% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGislet_annot.R
\name{CGI_annot}
\alias{CGI_annot}
\title{Get CGI annotation of CpG islets}
\usage{
CGI_annot(CG_islet, CGI)
}
\arguments{
\item{CG_islet}{CpG islet object}

\item{CGI}{CGI annotation file, bed format.}
}
\value{
a list includes CGI annotation of CpG islets
}
\description{
Get CGI annotation of CpG islets
}
