% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGislet.R
\name{CGislet_filter}
\alias{CGislet_filter}
\title{Filter CG islet candidates with pvalue}
\usage{
CGislet_filter(
  CGislet_object,
  genome,
  breaks = 10000,
  random_seq_num = 1e+05,
  CG_var = c("CpG_OE", "CpA_TpG_OE", "GC_content", "CG_density"),
  pvalueCutoff = 0.05,
  qvalueCutoff = 0.2
)
}
\arguments{
\item{genome}{Input genome sequences.}

\item{breaks}{CG_var of random DNA sequences were cut with breaks.}

\item{random_seq_num}{The number of random DNA sequences.}

\item{CG_var}{Features which were applied to filter CpG islets.}

\item{pvalueCutoff}{pvalueCutoff}

\item{qvalueCutoff}{qvalueCutoff}

\item{CGislet}{GRange object of CpG islets}
}
\value{
List includes filtered CpG islets range and CpG sites range
}
\description{
Filter CG islet candidates with pvalue
}
