% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGislet_annot.R
\name{CGislet_gene_annot}
\alias{CGislet_gene_annot}
\title{Annot CpG islets with gene annotation files}
\usage{
CGislet_gene_annot(
  CGislet,
  annot_file,
  genome,
  annotDb = NULL,
  annot_type = "gff3",
  tssRegion = c(-2500, 500)
)
}
\arguments{
\item{CGislet}{GRange object of CpG islets}

\item{annot_file}{path of annot file}

\item{genome}{Genome sequences, DNAStringSet object}

\item{annotDb}{Name of annotate database, org.HS.eg.db for human}

\item{annot_type}{Format of annot file, default: "gff3"}

\item{tssRegion}{Range of promoters}
}
\value{
CGislet with CGislet_type
}
\description{
Annot CpG islets with gene annotation files
}
