% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGislet.R
\name{CGislet_pathway_enrichment}
\alias{CGislet_pathway_enrichment}
\title{get type of CpG islets (core, linker, flank)}
\usage{
CGislet_pathway_enrichment(
  CGislet,
  annotDb = NULL,
  organism = NULL,
  GO = T,
  KEGG = T,
  GO_pvalueCutoff = 0.05,
  GO_qvalueCutoff = 0.1,
  KEGG_pvalueCutoff = 0.05,
  KEGG_qvalueCutoff = 0.1
)
}
\arguments{
\item{CGislet}{GRange object of CpG islets}

\item{annotDb}{Name of annotate database, org.HS.eg.db for human}

\item{organism}{organism name in KEGG database.}

\item{GO}{Logical, if T, do GO enrichment analysis.}

\item{KEGG}{Logical, if T, do KEGG enrichment analysis.}

\item{GO_pvalueCutoff}{GO enrichment pvalue cutoff}

\item{GO_qvalueCutoff}{GO enrichment qvalue cutoff}

\item{KEGG_pvalueCutoff}{KEGG enrichment pvalue cutoff}

\item{KEGG_qvalueCutoff}{KEGG enrichment qvalue cutoff}
}
\value{
GO/KEGG enrichment analysis
}
\description{
get type of CpG islets (core, linker, flank)
}
