% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGislet_annot.R
\name{get_annot}
\alias{get_annot}
\title{Get annotation from annot_list}
\usage{
get_annot(annot_list, pattern)
}
\arguments{
\item{annot_list}{annotation list split from GFF3 annotation column (V9)}

\item{pattern}{character string}
}
\description{
Get annotation from annot_list
}
