% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGislet.R
\name{get_CG_density}
\alias{get_CG_density}
\title{Calculate CG density in CpG range}
\usage{
get_CG_density(site, libs, flank_left = 500, flank_right = 502)
}
\arguments{
\item{site}{Genomic location of CG site.}

\item{libs}{Genomic locations of all CG site.}

\item{flank_left}{Length of left flank of CpG range. Default:500}

\item{flank_right}{Length of right flank of CpG range. Default:502}
}
\value{
CG density of input sequence.
}
\description{
Calculate CG density in CpG range
}
