% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGislet.R
\name{get_CG_dist_right}
\alias{get_CG_dist_right}
\title{Get dist of CpG site to nearest right CpG site}
\usage{
get_CG_dist_right(CG_start, genome_len)
}
\arguments{
\item{CG_start}{Start site of CpG sites.}

\item{genome_len}{Length of genome sequences.}
}
\value{
Dist of CpG site to nearest right CpG site
}
\description{
Get dist of CpG site to nearest right CpG site
}
