% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGislet.R
\name{get_CG_islet}
\alias{get_CG_islet}
\title{Get CG islet candidates}
\usage{
get_CG_islet(
  genome,
  flank_width = 500,
  Nmin = 10,
  Dmax = 75,
  prob = 0.6,
  Merge = T,
  Merge_length = 150
)
}
\arguments{
\item{genome}{Input genome sequences.}

\item{Nmin}{int. Minimal CpGs dinucleotide in CpG islets. Default:10}

\item{Dmax}{int. Maximal distance between 2 adjcent CpG islets. Default:75}

\item{prob}{numeric. CG site likelihood threshold, CG sites below this threshold will be filtered out. Default:0.6}

\item{Merge}{Logical. If TRUE, merge CpG islets candidates}

\item{Merge_length}{int. The CpG islets with a distance no more than M will be merged as a single CpG islet. Default:150}
}
\value{
List includes CpG islets range and CG sites range
}
\description{
Get CG islet candidates
}
