% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGislet.R
\name{get_CG_levels}
\alias{get_CG_levels}
\title{Cluster standardized CG sites based on their CpG_OE_ratio, CpA_TpG_OE_ratio, GC content, CG_density and dist with cmeans}
\usage{
get_CG_levels(
  CG_stat,
  CG_var = c("CpG_OE", "GC_content", "CpA_TpG_OE", "CG_density", "dist"),
  seed = 100,
  centers = 2,
  standard = T,
  order = T
)
}
\arguments{
\item{CG_stat}{GRanges object with CG stat information.}

\item{CG_var}{CpG_OE_ratio, CpA_TpG_OE_ratio, GC content, CG_density and dist}

\item{seed}{Seed of the random number generator.}

\item{centers}{number of cluster centroid.}

\item{standard}{Logical. Standardize CG sites or not. Default: TRUE.}

\item{order}{Logical. Order CG clusters or not. Default: TRUE.}

\item{filter}{Logical. Filter CG site with max dist. Default: TRUE.}

\item{max_dist}{Filter CG site with max dist. Default: 150.}
}
\value{
CG levels based on their CpG_OE_ratio, CpA_TpG_OE_ratio, CG_density and dist.
}
\description{
Cluster standardized CG sites based on their CpG_OE_ratio, CpA_TpG_OE_ratio, GC content, CG_density and dist with cmeans
}
