% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGislet.R
\name{get_CGislet_flank_site_pos}
\alias{get_CGislet_flank_site_pos}
\title{Detection of flank CpG site positions of each CpG islet}
\usage{
get_CGislet_flank_site_pos(CGislet, CG_stat, flank = 2000L, filter = T)
}
\arguments{
\item{CGislet}{CGislet objects contain CGislet_type column}

\item{flank}{Length of left and right flank of each CpG islet.Default: 2000L.}

\item{filter}{Logical. If TRUE, CpG sites in CpG islet were NOT considered as flank CpG sites.}

\item{CpG_site}{CG_site objects contain CG_islet column}
}
\value{
CG_site with flank CpG site positions
}
\description{
Detection of flank CpG site positions of each CpG islet
}
