% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGislet.R
\name{get_CGislet_seq}
\alias{get_CGislet_seq}
\title{Extract sequences of CG islets.}
\usage{
get_CGislet_seq(
  CG_islet,
  CG_site,
  genome,
  range = T,
  write_seq = FALSE,
  file_name = "CG_islet.fa",
  region = "all"
)
}
\arguments{
\item{CG_islet}{CG_islet object.}

\item{CG_site}{CG_site object.}

\item{genome}{Input genome sequences.}

\item{range}{If TRUE. Also return a genome range of choosed region. Valid if region in c("core","linker","flank")}

\item{write_seq}{if True,write CG islets sequences.Default:FALSE.}

\item{file_name}{filenames of CG islets sequences.Default:"CG_islet.fa".}

\item{region}{Obtain the sequence of the specified area in CGislets. One of "all","core","linker" or "flank"}
}
\value{
CG islets ranges with their CpG_OE_ratio CpA_TpG_OE_ratio CG_density and GC_content
}
\description{
Extract sequences of CG islets.
}
