% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGislet.R
\name{get_CGislet_type}
\alias{get_CGislet_type}
\title{get type of CpG islets (core, linker, flank)}
\usage{
get_CGislet_type(CGislet, CG_site)
}
\arguments{
\item{CGislet}{GRange object of CpG islets}

\item{CG_site}{GRange object of CpGs in CpG islets}
}
\value{
CGislet with CGislet_type
}
\description{
get type of CpG islets (core, linker, flank)
}
