% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGislet.R
\name{get_CpA_TpG_OE_ratio}
\alias{get_CpA_TpG_OE_ratio}
\title{get_CpA_TpG_OE_ratio}
\usage{
get_CpA_TpG_OE_ratio(sequences)
}
\arguments{
\item{sequences}{Input sequence.}
}
\value{
CpA_TpG_OE_ratio of input sequence
}
\description{
get_CpA_TpG_OE_ratio
}
\examples{
test_seq <- paste0(sample(c("A","T","C","G"),size = 100,replace = T),collapse = "")
get_CpA_TpG_OE_ratio(test_seq)
}
