% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGislet.R
\name{get_CpG_OE_ratio}
\alias{get_CpG_OE_ratio}
\title{Calculation of the CpG_OE_ratio of each sequences}
\usage{
get_CpG_OE_ratio(sequences)
}
\arguments{
\item{sequences}{Input sequence.}
}
\value{
CpG_OE_ratio of input sequence
}
\description{
Calculation of the CpG_OE_ratio of each sequences
}
\examples{
test_seq <- paste0(sample(c("A","T","C","G"),size = 100,replace = T),collapse = "")
get_CpG_OE_ratio(test_seq)
}
