% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGislet.R
\name{get_GC_content}
\alias{get_GC_content}
\title{Calculation of the GC content of each sequences}
\usage{
get_GC_content(sequence)
}
\arguments{
\item{sequences}{Input sequence.}
}
\value{
GC content of input sequence
}
\description{
Calculation of the GC content of each sequences
}
\examples{
test_seq <- paste0(sample(c("A","T","C","G"),size = 100,replace = T),collapse = "")
get_GC_content(test_seq)
}
