% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGislet.R
\name{get_pvalue}
\alias{get_pvalue}
\title{get pvalue with wilcox.test and permutation_test compared two group of features which were applied to filter CpG islets.}
\usage{
get_pvalue(
  dat,
  bootstrap = 100L,
  permutation = 1000L,
  vars = "CpG_OE",
  seed = 1234
)
}
\arguments{
\item{dat}{CG_islet list.}

\item{bootstrap}{Bootstrap sampling frequency.Default: 100.}

\item{permutation}{The number of permutations in permutation testing.}

\item{vars}{Features which were applied to filter CpG islets.}

\item{seed}{Seed of the random number generator.}
}
\value{
pvalue of two group of features
}
\description{
get pvalue with wilcox.test and permutation_test compared two group of features which were applied to filter CpG islets.
}
