% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGislet.R
\name{get_random_seqs}
\alias{get_random_seqs}
\title{get_random_seqs from genome}
\usage{
get_random_seqs(
  genome,
  random_seq_num = 100,
  random_seq_min = 200,
  random_seq_max = 800,
  cutoff = 100,
  seeds = 1234
)
}
\arguments{
\item{genome}{genome sequences (DNAStringSet).}

\item{random_seq_num}{Number of random sequences}

\item{random_seq_min}{Minimum length of random sequence.Default: 200}

\item{random_seq_max}{Maximum length of random sequence.Default: 800}

\item{cutoff}{Maximum number of non-DNA_BASES.Default: 100}

\item{seeds}{Random Number Seed.}
}
\value{
Random sequences extracted from imported genome
}
\description{
get_random_seqs from genome
}
