% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGislet_annot.R
\name{liftover}
\alias{liftover}
\title{Converts genome coordinates from result file of LiftOver}
\usage{
liftover(CGislet, liftover_file)
}
\arguments{
\item{CGislet}{GRange object including}

\item{liftover_file}{character;bed filepath obtained from LiftOver software}
}
\value{
position box ("chr4:100,001-100,001") of converted genome coordinates
}
\description{
Converts genome coordinates from result file of LiftOver
}
