% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGislet.R
\name{permutation_test}
\alias{permutation_test}
\title{Permutation test of features which were applied to filter CpG islets.}
\usage{
permutation_test(dat, n = 1000, vars = "CpG_OE")
}
\arguments{
\item{dat}{CG_islet list.}

\item{n}{The number of permutations in permutation testing.}

\item{vars}{Features which were applied to filter CpG islets.}
}
\value{
pvalue of permutation_test
}
\description{
Permutation test of features which were applied to filter CpG islets.
}
