% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGislet_annot.R
\name{range_annot}
\alias{range_annot}
\title{Get regulatory features annotation of CpG islets}
\usage{
range_annot(CG_islet, regulate_gff3)
}
\arguments{
\item{CG_islet}{CpG islet object}

\item{regulate_gff3}{Regulatory features annotation file; gff3 format}
}
\value{
a list includes regulatory features annotation of CpG islets
}
\description{
Get regulatory features annotation of CpG islets
}
