% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGislet.R
\name{search_CGislet}
\alias{search_CGislet}
\title{search CpG islet in genome}
\usage{
search_CGislet(
  CG_stat,
  Nmin,
  Dm,
  Mer,
  Merge_len,
  getseq = T,
  genome = genome,
  flank = 500,
  write_site_flank_seq = F,
  write_CpGislet_seq = F
)
}
\arguments{
\item{CG_stat}{Information of CpG sites.}

\item{Nmin}{int. Minimal CpGs dinucleotide in CpG islets. Default:10.}

\item{Dm}{int. Maximal distance between 2 adjcent CpG islets. Default:75.}

\item{Mer}{Logical. If true, CpG islets with dist lower than `Merge_len` would be merged.}

\item{Merge_len}{The CpG islets with a distance no more than M will be merged as a single CpG islet. Default:150.}

\item{genome}{genome sequences (DNAStringSet).}

\item{flank}{Flank width (bp) of each CpG site. Default: 500}

\item{write_site_flank_seq}{Logical. If true, write flank DNA sequence of each CpG site}

\item{write_CpGislet_seq}{Logical. If true, write CpG islet DNA sequences}
}
\value{
Information of CpG islets and CpG sites.
}
\description{
search CpG islet in genome
}
