% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGislet.R
\name{standard_CG}
\alias{standard_CG}
\title{Convert GC_content, CpG_OE_ratio, CpA_TpG_OE_ratio, CG_density and dist to nearest CG site
of each CG site flank sequences to uniform distribution}
\usage{
standard_CG(CG_mcol)
}
\arguments{
\item{CG_mcol}{Input parameters of CG sites (flank sequences).}
}
\value{
standardized CG sites
}
\description{
Convert GC_content, CpG_OE_ratio, CpA_TpG_OE_ratio, CG_density and dist to nearest CG site
of each CG site flank sequences to uniform distribution
}
