% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGislet_annot.R
\name{tsDMR_annot}
\alias{tsDMR_annot}
\title{Get tsDMR annotation of CpG islet}
\usage{
tsDMR_annot(CG_islet, tsDMR)
}
\arguments{
\item{CG_islet}{gff3 file path of data.frame}

\item{tsDMR}{tsDMR region GRange list}
}
\value{
a list includes tsDMR annotation of CpG islets
}
\description{
Get tsDMR annotation of CpG islet
}
